requirements.txt
numpy
matplotlib
pillow
opencv-python
pandas
scikit-learn
scikit-image
scipy
statsmodels
seaborn
statsmodels
statsmodels